Add private token method for System user Set main url in appsettings
Fix names:
AdviseToDoOfSubscription > SetUrlWebhook
Webhooks folder > HttpsClients

Fix gitignore
