+++ .gitignore
@@ -429,5 +429,3 @@ appsettings.Development.json
 
 # End of https://www.toptal.com/developers/gitignore/api/dotnetcore,visualstudio,visualstudiocode
 .DS_Store
-
-/src/Api/Controllers/WebhookController.cs
