+++ src/Api/appsettings.json
@@ -5,5 +5,8 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
+  "NotificationApi": {
+    "BaseUrl": "https://localhost:7057/"
+  },
   "AllowedHosts": "*"
 }
